Add files via upload
This code was used to generate the phase plane diagram. The Helfrich energy in combination with calculus of variations was used to derive the membrane shape equation for a GUV with two phases. A phase plane diagram was generated by finding solutions due to varying pressures and membrane tensions. Solutions were found using shoot-and-match methods, and initial conditions are found in the excel sheet.